requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0